feat: resolve easy circular dependencies
