+++ apps/bilan-carbone/prisma/seed/index.ts
@@ -4,7 +4,7 @@ import { reCreateBegesRules, reCreateGHGPRules } from '@/db/exports'
 import { getSectenVersion, updateSectenVersion } from '@/scripts/secten/secten'
 import { signPassword } from '@/services/auth'
 import { getEmissionFactorsFromAPI } from '@/services/importEmissionFactor/baseEmpreinte/getEmissionFactorsFromAPI'
-import { getAllowedLevels } from '@/services/study'
+import { getAllowedLevels } from '@/utils/study'
 import { faker } from '@faker-js/faker'
 import { PrismaPg } from '@prisma/adapter-pg'
 import type { Account, User } from '@repo/db-common'
