+++ apps/bilan-carbone/src/app/(public)/preview/etudes/[id]/ChartsPage.tsx
@@ -1,6 +1,6 @@
 import BarChart from '@/components/study/charts/BarChart'
 import PieChart from '@/components/study/charts/PieChart'
-import { BaseResultsByPost } from '@/services/results/consolidated'
+import type { BaseResultsByPost } from '@/services/posts'
 import { Translations } from '@/types/translation'
 import { StudyResultUnit } from '@repo/db-common/enums'
