+++ apps/bilan-carbone/src/components/study/infography/PostInfography.tsx
@@ -2,7 +2,7 @@
 
 import { BasePostInfography } from '@/environments/base/study/infography/BasePostInfography'
 import { Post, subPostsByPost } from '@/services/posts'
-import { ResultsByPost } from '@/services/results/consolidated'
+import { ResultsByPost } from '@/types/study.types'
 import { getEmissionValueString, getValidationPercentage } from '@/utils/study'
 import { StudyResultUnit, SubPost } from '@repo/db-common'
 import { useTranslations } from 'next-intl'
