+++ apps/bilan-carbone/src/components/study/results/ConsolidatedBEGESDifference.tsx
@@ -2,9 +2,9 @@ import { wasteEmissionFactors } from '@/constants/wasteEmissionFactors'
 import { EmissionFactorWithParts } from '@/db/emissionFactors'
 import type { FullStudy } from '@/db/study'
 import { Post } from '@/services/posts'
-import { ResultsByPost } from '@/services/results/consolidated'
 import { PostInfos } from '@/services/results/exports'
 import { getSiteEmissionSourcesWithoutMarketBase } from '@/services/results/utils'
+import { ResultsByPost } from '@/types/study.types'
 import { computeDifferenceForTableEmissions, formatDifferenceTableEmissions } from '@/utils/exports'
 import { formatNumber } from '@/utils/number'
 import { STUDY_UNIT_VALUES } from '@/utils/study'
