+++ apps/bilan-carbone/src/components/study/results/ConsolidatedExportDifference.tsx
@@ -3,8 +3,8 @@ import { EmissionFactorWithParts } from '@/db/emissionFactors'
 import type { FullStudy } from '@/db/study'
 import { customRich } from '@/i18n/customRich'
 import { getEmissionResults } from '@/services/emissionSource'
-import { ResultsByPost } from '@/services/results/consolidated'
 import { PostInfos } from '@/services/results/exports'
+import { ResultsByPost } from '@/types/study.types'
 import { formatNumber } from '@/utils/number'
 import { STUDY_UNIT_VALUES } from '@/utils/study'
 import LightbulbIcon from '@mui/icons-material/LightbulbOutlined'
