+++ apps/bilan-carbone/src/components/study/results/ResultFilters.tsx
@@ -1,6 +1,6 @@
 import type { FullStudy } from '@/db/study'
 import { environmentSubPostsMapping, Post, subPostsByPost } from '@/services/posts'
-import { ResultType } from '@/services/study'
+import type { ResultType } from '@/types/study.types'
 import { SubPost } from '@repo/db-common/enums'
 import { useEffect, useMemo, useRef, useState } from 'react'
 import { PostSubPostFilter } from '../../form/PostSubPostFilter'
