+++ apps/bilan-carbone/src/components/study/results/uncertainty/UncertaintyAnalytics.tsx
@@ -1,7 +1,7 @@
 import Title from '@/components/base/Title'
 import type { FullStudy } from '@/db/study'
-import { ResultsByPost } from '@/services/results/consolidated'
 import { getConfidenceInterval } from '@/services/uncertainty'
+import { ResultsByPost } from '@/types/study.types'
 import { Environment, StudyResultUnit } from '@repo/db-common/enums'
 import classNames from 'classnames'
 import { useTranslations } from 'next-intl'
