+++ apps/bilan-carbone/src/components/study/results/uncertainty/UncertaintyPerPost.tsx
@@ -3,8 +3,8 @@ import Title from '@/components/base/Title'
 import GlossaryModal from '@/components/modals/GlossaryModal'
 import { getAlpha } from '@/services/emissionSource'
 import { Post } from '@/services/posts'
-import { ResultsByPost } from '@/services/results/consolidated'
 import { getConfidenceInterval } from '@/services/uncertainty'
+import { ResultsByPost } from '@/types/study.types'
 import { formatEmissionFactorNumber, formatNumber } from '@/utils/number'
 import { defaultPostColor, postColors, STUDY_UNIT_VALUES } from '@/utils/study'
 import { ScatterMarkerProps, ScatterSeries } from '@mui/x-charts'
