+++ apps/bilan-carbone/src/components/study/rights/NewStudyRightForm.tsx
@@ -11,7 +11,7 @@ import { useServerFunction } from '@/hooks/useServerFunction'
 import { ALREADY_IN_STUDY } from '@/services/permissions/check'
 import { newStudyRight } from '@/services/serverFunctions/study'
 import { NewStudyRightCommand, NewStudyRightCommandValidation } from '@/services/serverFunctions/study.command'
-import { hasSufficientLevel } from '@/services/study'
+import { hasSufficientLevel } from '@/utils/study'
 import { isAdmin } from '@/utils/user'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { MenuItem } from '@mui/material'
