+++ apps/bilan-carbone/src/components/study/rights/StudyLevel.tsx
@@ -8,7 +8,7 @@ import { useServerFunction } from '@/hooks/useServerFunction'
 import { customRich } from '@/i18n/customRich'
 import { changeStudyLevel } from '@/services/serverFunctions/study'
 import { ChangeStudyLevelCommand, ChangeStudyLevelCommandValidation } from '@/services/serverFunctions/study.command'
-import { getAllowedLevels } from '@/services/study'
+import { getAllowedLevels } from '@/utils/study'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { MenuItem } from '@mui/material'
 import { Level } from '@repo/db-common/enums'
