+++ apps/bilan-carbone/src/components/study/site/useStudySite.ts
@@ -2,11 +2,10 @@
 
 import { storageKeys } from '@/constants/storage.constants'
 import type { FullStudy } from '@/db/study'
-import { StudyWithoutDetail } from '@/services/permissions/study'
 import { useRouter, useSearchParams } from 'next/navigation'
 import { useEffect, useMemo, useRef, useState } from 'react'
 
-export default function useStudySite(study: FullStudy | StudyWithoutDetail, allowAll?: boolean) {
+export default function useStudySite(study: FullStudy, allowAll?: boolean) {
   const [siteId, setSiteState] = useState('')
   const searchParams = useSearchParams()
   const router = useRouter()
