+++ apps/bilan-carbone/src/db/study.ts
@@ -5,10 +5,9 @@ import { hasAccessToCreateStudyWithEmissionFactorVersions } from '@/services/per
 import { filterAllowedStudies } from '@/services/permissions/study'
 import { Post, subPostsByPost } from '@/services/posts'
 import { ChangeStudyCinemaCommand } from '@/services/serverFunctions/study.command'
-import { getAllowedLevels, hasSufficientLevel } from '@/services/study'
 import { mapCncToStudySite } from '@/utils/cnc'
 import { isAdminOnOrga } from '@/utils/organization'
-import { getUserRoleOnPublicStudy, StudyWithRoleFields } from '@/utils/study'
+import { getAllowedLevels, getUserRoleOnPublicStudy, hasSufficientLevel, StudyWithRoleFields } from '@/utils/study'
 import { isAdmin } from '@/utils/user'
 import type { DuplicableStudy, Level, Prisma, StudyTag, StudyTagFamily, SubPost } from '@repo/db-common'
 import { CommentStatus, ControlMode, Environment, Export, Import, StudyRole } from '@repo/db-common/enums'
