+++ apps/bilan-carbone/src/environments/base/study/new/Form.tsx
@@ -12,7 +12,7 @@ import StudyExportsForm from '@/components/study/perimeter/StudyExportsForm'
 import { getOrganizationVersionAccounts } from '@/db/organization'
 import type { FullStudy } from '@/db/study'
 import { CreateStudyCommand } from '@/services/serverFunctions/study.command'
-import { getAllowedLevels, hasSufficientLevel } from '@/services/study'
+import { getAllowedLevels, hasSufficientLevel } from '@/utils/study'
 import { FormControlLabel, MenuItem, Radio } from '@mui/material'
 import { Level } from '@repo/db-common/enums'
 import { UserSession } from 'next-auth'
