+++ apps/bilan-carbone/src/environments/clickson/study/PDF/PDFSummary.tsx
@@ -5,7 +5,7 @@ import '@/app/(public)/preview/etudes/[id]/pdf-summary.css'
 import ConsolidatedResultsTable from '@/components/study/results/consolidated/ConsolidatedResultsTable'
 import type { FullStudy } from '@/db/study'
 import { usePublicodesResults } from '@/hooks/usePublicodesResults'
-import { BaseResultsByPost } from '@/services/results/consolidated'
+import type { BaseResultsByPost } from '@/services/posts'
 import { ThemeProvider } from '@mui/material/styles'
 import { useTranslations } from 'next-intl'
 import Image from 'next/image'
