+++ apps/bilan-carbone/src/environments/cut/study/PDF/PDFSummary.tsx
@@ -6,8 +6,8 @@ import ConsolidatedResultsTable from '@/components/study/results/consolidated/Co
 import type { FullStudy } from '@/db/study'
 import cutTheme from '@/environments/cut/theme/theme'
 import { usePublicodesResults } from '@/hooks/usePublicodesResults'
+import type { BaseResultsByPost } from '@/services/posts'
 import { convertSimplifiedEnvToBilanCarbone } from '@/services/posts'
-import { BaseResultsByPost } from '@/services/results/consolidated'
 import { getTotalValueFromBaseResults } from '@/services/results/publicodes'
 import { formatNumber } from '@/utils/number'
 import { ThemeProvider } from '@mui/material/styles'
