+++ apps/bilan-carbone/src/environments/simplified/study/infography/AllPostsInfography.tsx
@@ -3,8 +3,8 @@ import EnvironmentLoader from '@/environments/core/utils/EnvironmentLoader'
 import { usePublicodesSituation } from '@/lib/publicodes/context'
 import { mappedTiltSituationToCustomDataFields } from '@/services/customDataToSituation'
 import { isTilt } from '@/services/permissions/environment'
+import type { BaseResultsByPost } from '@/services/posts'
 import { getQuestionProgressBySubPost, StatsResult } from '@/services/publicodes/questionProgress'
-import { BaseResultsByPost } from '@/services/results/consolidated'
 import { computeBaseResultsByPostFromEngine } from '@/services/results/publicodes'
 import { useAppEnvironmentStore } from '@/store/AppEnvironment'
 import { getEmissionValueString } from '@/utils/study'
