+++ apps/bilan-carbone/src/environments/tilt/study/infography/AllPostsInfography.tsx
@@ -3,7 +3,7 @@ import PostInfography from '@/components/study/infography/PostInfography'
 import type { FullStudy } from '@/db/study'
 import theme from '@/environments/tilt/theme/theme'
 import { TiltPost } from '@/services/posts'
-import { ResultsByPost } from '@/services/results/consolidated'
+import { ResultsByPost } from '@/types/study.types'
 import { Box } from '@mui/material'
 import classNames from 'classnames'
 import { useTranslations } from 'next-intl'
