+++ apps/bilan-carbone/src/services/study.utils.ts
@@ -2,12 +2,13 @@ import { EngagementActionTargets } from '@/constants/engagementActions'
 import type { FullStudy } from '@/db/study'
 import { Translations } from '@/types/translation'
 import { getEmissionFactorValue } from '@/utils/emissionFactors'
+import { getEmissionSourcesTotalCo2 } from '@/utils/emissionSources'
 import { getPost } from '@/utils/post'
 import { formatEmissionValueForExport, hasDeprecationPeriod, isCAS } from '@/utils/study'
 import { formatDateFr } from '@/utils/time'
 import { Environment, Level, StudyResultUnit, SubPost } from '@repo/db-common/enums'
 import dayjs from 'dayjs'
-import { getEmissionResults, getEmissionSourceEmission, getEmissionSourcesTotalCo2 } from './emissionSource'
+import { getEmissionResults, getEmissionSourceEmission } from './emissionSource'
 import { download } from './file'
 import { Post } from './posts'
 import { EmissionFactorWithMetaData, getEmissionFactorsByIds } from './serverFunctions/emissionFactor'
