+++ apps/bilan-carbone/src/types/emissionSource.types.ts
@@ -1,0 +1,6 @@
+export enum EmissionSourcesStatus {
+  Valid = 'valid',
+  ToVerify = 'toVerify',
+  Waiting = 'waiting',
+  WaitingContributor = 'waitingContributor',
+}
